move handleDuplicateSpendingInfoDb into DLCWalletLoaderApi
